Refactor : Move hip::HipTraceManager to its own translation unit
